# Conflicts:
#	app/src/main/java/com/jeju/nanaland/ui/navigation/MainScreenRoute.kt

+++ app/src/main/java/com/jeju/nanaland/domain/entity/search/SearchResultData.kt
@@ -4,7 +4,7 @@ import com.google.gson.annotations.SerializedName
 
 data class SearchResultData(
     @SerializedName("totalElements")
-    val count: Long,
+    val count: Int,
     @SerializedName("data")
     val data: List<SearchResultThumbnailData>
 )
