+++ app/src/main/java/com/jeju/nanaland/util/intent/DeepLinkData.kt
@@ -3,5 +3,5 @@ package com.jeju.nanaland.util.intent
 data class DeepLinkData(
     var language: String? = null,
     var category: String? = null,
-    var contentId: Long? = null
+    var contentId: Int? = null
 )
